+++ Modules/DataModel/Sources/PocketCastsDataModel/Private/DB/PCDatabase.swift
@@ -3,6 +3,8 @@ import Foundation
 public protocol PCDatabase {
     var changes: Int32 { get }
 
+    func pragmaUserVersion() -> Int32?
+
     func executeQuery(_ sql: String, values: [Any]?) throws -> PCDBResultSet
 
     func executeUpdate(_ sql: String, values: [Any]?) throws