+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 7.89
 -----
-
+- Add Recommendations feature to Discover and Podcast pages [#2989](https://github.com/Automattic/pocket-casts-ios/issues/2989)
 
 7.88
 -----